Tests/LibWeb/Text/input/HTML/script-async-attribute.html

<!DOCTYPE html>
<script src="../include.js" id="include-script"></script>
<script>
    test(() => {
        const includeScript = document.querySelector("#include-script");
        println(`includeScript.async = ${includeScript.async}`);

        const script = document.createElement("script");
        println(`script.async = ${script.async}`);
    });
</script>
