Tests/LibWeb/Text/expected/HTML/script-async-attribute.txt

includeScript.async = false
script.async = true
